requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
crcmod==1.7
Deprecated==1.3.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
future==1.0.0
greenlet==3.3.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
line-bot-sdk==3.21.0
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
pillow==12.1.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
qrcode==8.2
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wrapt==2.0.1
yarl==1.22.0
